# kv: optimize IsPoint