--- a/kv/key_test.go
+++ b/kv/key_test.go
@@ -15,6 +15,7 @@ package kv
 
 import (
 	"bytes"
+	"testing"
 	"time"
 
 	. "github.com/pingcap/check"
@@ -55,18 +56,64 @@ func (s *testKeySuite) TestPartialNext(c *C) {
 }
 
 func (s *testKeySuite) TestIsPoint(c *C) {
+	tests := []struct {
+		start   []byte
+		end     []byte
+		isPoint bool
+	}{
+		{
+			start:   Key("rowkey1"),
+			end:     Key("rowkey2"),
+			isPoint: true,
+		},
+		{
+			start:   Key("rowkey1"),
+			end:     Key("rowkey3"),
+			isPoint: false,
+		},
+		{
+			start:   Key(""),
+			end:     []byte{0},
+			isPoint: true,
+		},
+		{
+			start:   []byte{123, 123, 255, 255},
+			end:     []byte{123, 124, 0, 0},
+			isPoint: true,
+		},
+		{
+			start:   []byte{123, 123, 255, 255},
+			end:     []byte{123, 124, 0, 1},
+			isPoint: false,
+		},
+	}
+	for _, tt := range tests {
+		kr := KeyRange{
+			StartKey: tt.start,
+			EndKey:   tt.end,
+		}
+		c.Check(kr.IsPoint(), Equals, tt.isPoint)
+	}
+}
+
+func BenchmarkIsPoint(b *testing.B) {
+	b.ReportAllocs()
 	kr := KeyRange{
-		StartKey: Key("rowkey1"),
-		EndKey:   Key("rowkey2"),
+		StartKey: []byte("rowkey1"),
+		EndKey:   []byte("rowkey2"),
 	}
-	c.Check(kr.IsPoint(), IsTrue)
-
-	kr.EndKey = Key("rowkey3")
-	c.Check(kr.IsPoint(), IsFalse)
-
-	kr = KeyRange{
-		StartKey: Key(""),
-		EndKey:   Key([]byte{0}),
+	for i := 0; i < b.N; i++ {
+		kr.IsPoint()
 	}
-	c.Check(kr.IsPoint(), IsTrue)
+}
+
+func BenchmarkIsPointByPrefixNext(b *testing.B) {
+	b.ReportAllocs()
+	kr := KeyRange{
+		StartKey: []byte("rowkey1"),
+		EndKey:   []byte("rowkey2"),
+	}
+	for i := 0; i < b.N; i++ {
+		bytes.Equal(kr.StartKey.PrefixNext(), kr.EndKey)
+	}
 }