<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
Improve performance of `IsPoint`.
```
benchmark               old ns/op     new ns/op     delta
BenchmarkIsPoint-12     5.35          21.1          +294.39%

benchmark               old allocs     new allocs     delta
BenchmarkIsPoint-12     0              1              +Inf%

benchmark               old bytes     new bytes     delta
BenchmarkIsPoint-12     0             8             +Inf%
```

### What is changed and how it works?
Avoid memory allocation by comparing bytes directly.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test